review comments - Renamed DDLParse -> DDLProcessor
