+++ yb-voyager/src/queryparser/query_parser.go
@@ -19,9 +19,7 @@ import (
 	"fmt"
 
 	pg_query "github.com/pganalyze/pg_query_go/v5"
-	"github.com/samber/lo"
 	log "github.com/sirupsen/logrus"
-	"google.golang.org/protobuf/reflect/protoreflect"
 )
 
 func Parse(query string) (*pg_query.ParseResult, error) {
@@ -42,250 +40,21 @@ func ParsePLPGSQLToJson(query string) (string, error) {
 	return jsonString, err
 }
 
-func ParseDDL(query string) (DDLObject, error) {
+func ParseAndProcessDDL(query string) (DDLObject, error) {
 	parseTree, err := Parse(query)
 	if err != nil {
 		return nil, fmt.Errorf("parsing query failed: %v", err)
 	}
 
-	parser, err := GetDDLParser(parseTree)
+	processor, err := GetDDLProcessor(parseTree)
 	if err != nil {
-		return nil, fmt.Errorf("getting parser failed: %v", err)
+		return nil, fmt.Errorf("getting processor failed: %v", err)
 	}
 
-	ddlObject, err := parser.Parse(parseTree)
+	ddlObject, err := processor.Process(parseTree)
 	if err != nil {
 		return nil, fmt.Errorf("parsing DDL failed: %v", err)
 	}
 
 	return ddlObject, nil
-}
-
-func DeparseSelectStmt(selectStmt *pg_query.SelectStmt) (string, error) {
-	if selectStmt != nil {
-		parseResult := &pg_query.ParseResult{
-			Stmts: []*pg_query.RawStmt{
-				{
-					Stmt: &pg_query.Node{
-						Node: &pg_query.Node_SelectStmt{SelectStmt: selectStmt},
-					},
-				},
-			},
-		}
-
-		// Deparse the SelectStmt to get the string representation
-		selectSQL, err := pg_query.Deparse(parseResult)
-		return selectSQL, err
-	}
-	return "", nil
-}
-
-func GetProtoMessageFromParseTree(parseTree *pg_query.ParseResult) protoreflect.Message {
-	return parseTree.Stmts[0].Stmt.ProtoReflect()
-}
-
-func IsPLPGSQLObject(parseTree *pg_query.ParseResult) bool {
-	// CREATE FUNCTION is same parser NODE for FUNCTION/PROCEDURE
-	_, isPlPgSQLObject := getCreateFuncStmtNode(parseTree)
-	return isPlPgSQLObject
-}
-
-func IsViewObject(parseTree *pg_query.ParseResult) bool {
-	_, isViewStmt := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_ViewStmt)
-	return isViewStmt
-}
-
-func IsMviewObject(parseTree *pg_query.ParseResult) bool {
-	createAsNode, isCreateAsStmt := getCreateTableAsStmtNode(parseTree) //for MVIEW case
-	return isCreateAsStmt && createAsNode.CreateTableAsStmt.Objtype == pg_query.ObjectType_OBJECT_MATVIEW
-}
-
-func GetSelectStmtQueryFromViewOrMView(parseTree *pg_query.ParseResult) (string, error) {
-	viewNode, isViewStmt := getCreateViewNode(parseTree)
-	createAsNode, _ := getCreateTableAsStmtNode(parseTree) //For MVIEW case
-	var selectStmt *pg_query.SelectStmt
-	if isViewStmt {
-		selectStmt = viewNode.ViewStmt.GetQuery().GetSelectStmt()
-	} else {
-		selectStmt = createAsNode.CreateTableAsStmt.GetQuery().GetSelectStmt()
-	}
-	selectStmtQuery, err := DeparseSelectStmt(selectStmt)
-	if err != nil {
-		return "", fmt.Errorf("deparsing the select stmt: %v", err)
-	}
-	return selectStmtQuery, nil
-}
-
-func GetObjectTypeAndObjectName(parseTree *pg_query.ParseResult) (string, string) {
-	createFuncNode, isCreateFunc := getCreateFuncStmtNode(parseTree)
-	viewNode, isViewStmt := getCreateViewNode(parseTree)
-	createAsNode, _ := getCreateTableAsStmtNode(parseTree)
-	createTableNode, isCreateTable := getCreateTableStmtNode(parseTree)
-	createIndexNode, isCreateIndex := getCreateIndexStmtNode(parseTree)
-	alterTableNode, isAlterTable := getAlterStmtNode(parseTree)
-	switch true {
-	case isCreateFunc:
-		/*
-			version:160001 stmts:{stmt:{create_function_stmt:{replace:true funcname:{string:{sval:"public"}} funcname:{string:{sval:"add_employee"}}
-			parameters:{function_parameter:{name:"emp_name" arg_type:{names:{string:{sval:"pg_catalog"}} names:{string:{sval:"varchar"}}
-			typemod:-1 location:62} mode:FUNC_PARAM_DEFAULT}} parameters:{funct ...
-
-			version:160001 stmts:{stmt:{create_function_stmt:{is_procedure:true replace:true funcname:{string:{sval:"public"}}
-			funcname:{string:{sval:"add_employee"}} parameters:{function_parameter:{name:"emp_name" arg_type:{names:{string:{sval:"pg_catalog"}}
-			names:{string:{sval:"varchar"}} typemod:-1 location:63} mode:FUNC_PARAM_DEFAULT}} ...
-		*/
-		stmt := createFuncNode.CreateFunctionStmt
-		objectType := "FUNCTION"
-		if stmt.IsProcedure {
-			objectType = "PROCEDURE"
-		}
-		funcNameList := stmt.GetFuncname()
-		return objectType, getFunctionObjectName(funcNameList)
-	case isViewStmt:
-		viewName := viewNode.ViewStmt.View
-		return "VIEW", getObjectNameFromRangeVar(viewName)
-	case IsMviewObject(parseTree):
-		intoMview := createAsNode.CreateTableAsStmt.Into.Rel
-		return "MVIEW", getObjectNameFromRangeVar(intoMview)
-	case isCreateTable:
-		return "TABLE", getObjectNameFromRangeVar(createTableNode.CreateStmt.Relation)
-	case isAlterTable:
-		return "TABLE", getObjectNameFromRangeVar(alterTableNode.AlterTableStmt.Relation)
-	case isCreateIndex:
-		indexName := createIndexNode.IndexStmt.Idxname
-		schemaName := createIndexNode.IndexStmt.Relation.GetSchemaname()
-		tableName := createIndexNode.IndexStmt.Relation.GetRelname()
-		fullyQualifiedName := lo.Ternary(schemaName != "", schemaName+"."+tableName, tableName)
-		displayObjName := fmt.Sprintf("%s ON %s", indexName, fullyQualifiedName)
-		return "INDEX", displayObjName
-	default:
-		panic("unsupported type of parseResult")
-	}
-}
-
-func GetIndexAccessMethod(parseTree *pg_query.ParseResult) string {
-	indexNode, _ := getCreateIndexStmtNode(parseTree)
-	return indexNode.IndexStmt.AccessMethod
-}
-
-// Range Var is the struct to get the relation information like relation name, schema name, persisted relation or not, etc..
-func getObjectNameFromRangeVar(obj *pg_query.RangeVar) string {
-	schema := obj.Schemaname
-	name := obj.Relname
-	return lo.Ternary(schema != "", fmt.Sprintf("%s.%s", schema, name), name)
-}
-
-func getFunctionObjectName(funcNameList []*pg_query.Node) string {
-	funcName := ""
-	funcSchemaName := ""
-	if len(funcNameList) > 0 {
-		funcName = funcNameList[len(funcNameList)-1].GetString_().Sval // func name can be qualified / unqualifed or native / non-native proper func name will always be available at last index
-	}
-	if len(funcNameList) >= 2 { // Names list will have all the parts of qualified func name
-		funcSchemaName = funcNameList[len(funcNameList)-2].GetString_().Sval // // func name can be qualified / unqualifed or native / non-native proper schema name will always be available at last 2nd index
-	}
-	return lo.Ternary(funcSchemaName != "", fmt.Sprintf("%s.%s", funcSchemaName, funcName), funcName)
-}
-
-func getTypeNameAndSchema(typeNames []*pg_query.Node) (string, string) {
-	typeName := ""
-	typeSchemaName := ""
-	if len(typeNames) > 0 {
-		typeName = typeNames[len(typeNames)-1].GetString_().Sval // type name can be qualified / unqualifed or native / non-native proper type name will always be available at last index
-	}
-	if len(typeNames) >= 2 { // Names list will have all the parts of qualified type name
-		typeSchemaName = typeNames[len(typeNames)-2].GetString_().Sval // // type name can be qualified / unqualifed or native / non-native proper schema name will always be available at last 2nd index
-	}
-
-	return typeName, typeSchemaName
-}
-
-func getCreateTableAsStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreateTableAsStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreateTableAsStmt)
-	return node, ok
-}
-
-func getCreateViewNode(parseTree *pg_query.ParseResult) (*pg_query.Node_ViewStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_ViewStmt)
-	return node, ok
-}
-
-func getCreateFuncStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreateFunctionStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreateFunctionStmt)
-	return node, ok
-}
-
-func getCreateTableStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreateStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreateStmt)
-	return node, ok
-}
-
-func getCreateIndexStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_IndexStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_IndexStmt)
-	return node, ok
-}
-
-func getAlterStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_AlterTableStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_AlterTableStmt)
-	return node, ok
-}
-
-func getCreateTriggerStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreateTrigStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreateTrigStmt)
-	return node, ok
-}
-
-func getPolicyStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreatePolicyStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreatePolicyStmt)
-	return node, ok
-}
-
-func getCompositeTypeStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CompositeTypeStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CompositeTypeStmt)
-	return node, ok
-}
-
-func getEnumTypeStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreateEnumStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreateEnumStmt)
-	return node, ok
-}
-func getForeignTableStmtNode(parseTree *pg_query.ParseResult) (*pg_query.Node_CreateForeignTableStmt, bool) {
-	node, ok := parseTree.Stmts[0].Stmt.Node.(*pg_query.Node_CreateForeignTableStmt)
-	return node, ok
-}
-
-func IsAlterTable(parseTree *pg_query.ParseResult) bool {
-	_, isAlter := getAlterStmtNode(parseTree)
-	return isAlter
-}
-
-func IsCreateIndex(parseTree *pg_query.ParseResult) bool {
-	_, isCreateIndex := getCreateIndexStmtNode(parseTree)
-	return isCreateIndex
-}
-
-func IsCreateTable(parseTree *pg_query.ParseResult) bool {
-	_, IsCreateTable := getCreateTableStmtNode(parseTree)
-	return IsCreateTable
-}
-
-func IsCreatePolicy(parseTree *pg_query.ParseResult) bool {
-	_, isCreatePolicy := getPolicyStmtNode(parseTree)
-	return isCreatePolicy
-}
-
-func IsCreateTrigger(parseTree *pg_query.ParseResult) bool {
-	_, isCreateTrig := getCreateTriggerStmtNode(parseTree)
-	return isCreateTrig
-}
-
-func IsCreateType(parseTree *pg_query.ParseResult) bool {
-	_, isComposite := getCompositeTypeStmtNode(parseTree)
-	_, isEnum := getEnumTypeStmtNode(parseTree)
-	return isComposite || isEnum
-}
-
-func IsCreateForeign(parseTree *pg_query.ParseResult) bool {
-	_, isForeignTable := getForeignTableStmtNode(parseTree)
-	return isForeignTable
 }
